pyproject.toml
[tool.poetry]
name = "harleen20"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
praw = "^7.4.0"
pyjokes = "^0.6.0"
requests = "^2.26.0"
wikipedia = "^1.4.0"
"randomstuff.py" = "^2.0.0"
bs4 = "^0.0.1"
Flask = "^2.0.1"
python-dotenv = "^0.19.0"
youtube-search = "^2.1.0"
"googlesearch.py" = "^1.2"
TenGiphPy = "^2.5.0"
youtube_dl = "^2021.6.6"
PyNaCl = "^1.4.0"
google-api-python-client = "^2.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
